feat: 입력 장치 선택하기 헤더 모달에 연결 #67
입력 장치 선택을 헤더의 모달에서 선택할 수 있게 했습니다. 선택값은
recoil을 이용하여 저장하였습니다.
