+++ frontend/src/components/Header/components/HeaderSettingModal.tsx
@@ -1,9 +1,29 @@
+import { useState, useEffect, useRef } from "react";
+import { useSetRecoilState } from "recoil";
+import selectedMicrophoneState from "./MicState";
+
 interface HeaderSettingModalProps {
   isSettingClicked: boolean;
   setIsSettingClicked: React.Dispatch<React.SetStateAction<boolean>>;
 }
 
 const HeaderSettingModal = ({ isSettingClicked, setIsSettingClicked }: HeaderSettingModalProps) => {
+  const [microphoneDevices, setMicrophoneDevices] = useState<MediaDeviceInfo[]>([]);
+  const setSelectedMicrophone = useSetRecoilState(selectedMicrophoneState);
+
+  useEffect(() => {
+    // 마이크 장치 목록 가져오기
+    navigator.mediaDevices
+      .enumerateDevices()
+      .then((devices) => {
+        const microphones = devices.filter((device) => device.kind === "audioinput");
+        setMicrophoneDevices(microphones);
+      })
+      .catch((error) => {
+        console.error("입력 장치 목록 불러오기 실패", error);
+      });
+  }, []);
+
   return (
     <>
       <div
@@ -20,6 +40,13 @@ const HeaderSettingModal = ({ isSettingClicked, setIsSettingClicked }: HeaderSet
         <div className="flex flex-row gap-3 w-full h-14 justify-start">
           <p className="semibold-18">입력 장치 설정</p>
         </div>
+
+        <select className="border semibold-18" onChange={(e) => setSelectedMicrophone(e.target.value)}>
+          {microphoneDevices.map((device) => (
+            <option value={device.deviceId}>{device.label || `마이크 ${microphoneDevices.indexOf(device) + 1}`}</option>
+          ))}
+        </select>
+
         <div className="flex flex-row gap-4 w-full"></div>
       </div>
     </>
