+++ frontend/src/components/Header/components/MicState.tsx
@@ -1,0 +1,9 @@
+import { atom } from "recoil";
+
+export const selectedMicrophoneState = atom<string>({
+  key: "selectedMicrophone",
+  default: "default",
+  dangerouslyAllowMutability: true
+});
+
+export default selectedMicrophoneState;
